requirement.txt
fastapi
uvicorn[standard]
pydantic
requests
demjson3
python-pptx
python-dotenv
groq
pymongo
